+++ one/params.py
@@ -8,6 +8,7 @@ directories, and a separate parameter file for each url containing the client pa
 caches file also sets the default client for when no url is provided.
 
 TODO Rename 'client' kwarg
+FIXME Remove redundant cache_dir param (use directory from client map instead)
 """
 import re
 from iblutil.io import params as iopar
@@ -66,6 +67,7 @@ def setup(client=None, silent=False, make_default=None):
     # If a client URL has been provided, set it as the default URL
     par_default = par_default.set('ALYX_URL', client or par_default.ALYX_URL)
     par_current = iopar.read(f'{_PAR_ID_STR}/{client_key}', par_default)
+
     # Load the db URL map
     cache_map = iopar.read(f'{_PAR_ID_STR}/{_CLIENT_ID_STR}', {'CLIENT_MAP': dict()})
     cache_dir = cache_map.CLIENT_MAP.get(client_key, Path(CACHE_DIR_DEFAULT, client_key))
@@ -136,10 +138,13 @@ def get(silent=False, client=None):
 
 
 def save(par, client):
+    # Remove cache dir variable before saving
+    par = {k: v for k, v in par.as_dict().items() if 'CACHE_DIR' not in k}
     iopar.write(f'{_PAR_ID_STR}/{_key_from_url(client)}', par)
 
 
 def get_cache_dir() -> Path:
+    # TODO Add client param
     cache_map = iopar.read(f'{_PAR_ID_STR}/{_CLIENT_ID_STR}', {})
     cache_dir = Path(cache_map.CLIENT_MAP[cache_map.DEFAULT] if cache_map else CACHE_DIR_DEFAULT)
     cache_dir.mkdir(exist_ok=True, parents=True)